Simplify configuration to one file
